djangocms_layout/tests/requirements.txt
# requirements from setup.py
# other requirements
djangocms-helper>=0.9.2
tox
coverage
